circuits: zk-gadgets: poseidon: Rearrange into module
